MINIST/requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.1.0
Pillow>=9.3.0
flask>=2.3.0
